Only update readline
This host library version ends up in the new stage1, however as part of
a stage1, readline is updated.

In this instance, readline 6 is on the host and readline 7 is installed
into the stage1, but bash (which is pretty important) is built
againstthe hosts' readline version (in this case, 6).

Update readline in the seed to upgrade it to version 7.